post index view, post edit

+++ app/Post.php
@@ -2,6 +2,7 @@
 
 namespace App;
 
+use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Model;
 use Cviebrock\EloquentSluggable\Sluggable;
 use Illuminate\Support\Facades\Storage;
@@ -15,14 +16,14 @@ class Post extends Model
     const IS_PUBLIC = 1;
 
 
-    protected $fillable = ['title', 'content'];
+    protected $fillable = ['title', 'content', 'date'];
 
     public function category(){
-        return $this->hasOne(Category::class);
+        return $this->belongsTo(Category::class, 'category_id');
     }
 
     public function author(){
-        return $this->hasOne(User::class);
+        return $this->belongsTo(User::class, 'user_id');
     }
 
     public function tags(){
@@ -57,17 +58,23 @@ class Post extends Model
     }
 
     public function remove(){
-        Storage::delete('uploads/' . $this->image);
+        $this->removeImage();
         $this->delete();
+    }
+
+    public function removeImage(){ //проверка на наличие и удаление картинки
+        if($this->image != null) {
+            Storage::delete('uploads/' . $this->image);
+        }
     }
 
     public function uploadImage($image){
         if($image == null) { //если картинка не зашла - выйти
             return;
         }
-        Storage::delete('uploads/' . $this->image);
+        $this->removeImage();
         $filename = str_random(10) . '.' . $image->extension();
-        $image->saveAS('uploads', '$filename');
+        $image->storeAS('uploads', $filename);
         $this->image = $filename;
         $this->save();
     }
@@ -128,6 +135,23 @@ class Post extends Model
         return '/uploads/' . $this->image;
     }
 
+    public function setDateAttribute($value) {
+        $date = Carbon::createFromFormat('d/m/y', $value)->format('Y-m-d');
+        $this->attributes['date'] = $date;
+//        echo date("H:i:s");
+    }
+
+    public function getCategoryTitle() {
+        if($this->category != null) {
+            return $this->category->title;
+        } else {
+            return 'Нет категории';
+        }
+    }
+
+    public function getTagsTitles() {
+        dd($this->tags);
+    }
 
 
 }
